test(task-migrate-clean-batch): 迁 overlay_labels 的 38 个用例到 gui_unit_test（清单 37→38）
420.5 Step 12，本批单文件迁移量最大的一步。

**与 420.3 清单「37/39」的偏差：实际可迁 38 个**
被 420.3 排除的 single_orthographic_dispatch_matches_fisheye 其实完全是纯逻辑。
它被漏掉是因为注册形态是 `t->TestFunc = SingleOrthoLatTestFunc;`（具名函数而非
lambda），静态扫描找不到 lambda 里的 IM_UNUSED(ctx)——而那个函数体第一行就是
IM_UNUSED(ctx)。这与 Step 10 的两个假阳性同源、方向相反：`IM_UNUSED(ctx)` 这条
静态标记两个方向都不可靠，只能读函数体。

真正需要活帧的只有 modal_does_not_leak_to_foreground 一个：它断言的是
DrawOverlayLabels 写进哪张 draw list（窗口 vs 前景），而 draw list 只在帧内存在。

- 新增 test/unit-correctness/gui/test_overlay_labels.cpp（38 个 TEST：
  OverlayLabels 32 + OverlayToggle 6，按原 ImGui category 分两个 suite）
  五个 helper（MakeGridOnly / MakeSunOnly / Count*Labels）随之迁走
- 原文件删 38 个块（39→1），头注释重写，空转的 helper 与 include 清掉
- cc1-cc4 四个 doc/overlay-label-placement.md 缺口回归锚点逐条核对，断言语义未弱化

顺带修一个迁移工具盲区（已在两个脚本里补上）：`IM_CHECK_GT_NO_RET` /
`IM_CHECK_LT_NO_RET` 不被 `\bIM_CHECK_GT\b` 匹配（下划线是词内字符），
首轮转换漏掉 7 处，编译期以 undeclared identifier 暴露。

核验：源文件侧断言减量 IM_CHECK 12 + IM_CHECK_EQ 33 + IM_CHECK_GE 7 +
IM_CHECK_GT 13 + IM_CHECK_GT_NO_RET 3 + IM_CHECK_LE 3 + IM_CHECK_LT 18 +
IM_CHECK_LT_NO_RET 4 = 93，新文件 EXPECT_TRUE 12 / EXPECT_EQ 33 / EXPECT_GE 7 /
EXPECT_GT 16 / EXPECT_LE 3 / EXPECT_LT 22 = 93，逐类对上；
gui_test 静态用例数 479→441；gui_unit_test 96→134 全绿。
